Avoid nested class that leads to NI trouble
